pyproject.toml
[tool.poetry]
name = "neuralab"
version = "0.1.0"
description = ""
authors = ["J.David Luque <jdluque@leitat.org>"]
readme = "README.md"

[tool.poetry.scripts]
dataframe = "neuralab.trading.dataframe:cli"
dataset = "neuralab.trading.dataset:cli"
nl = "neuralab.__main__:cli"

[tool.poetry.dependencies]
python = "^3.12"
#flax = "^0.10.2"
torch = {extras = ["cpu"], version = "^2.5.1"}
pandas = "^2.2.3"
matplotlib = "^3.9.3"
einops = "^0.8.0"
treescope = "^0.1.7"
#jaxtyping = "^0.2.36"
requests-toolbelt = "^1.0.0"
atpbar = {extras = ["jupyter"], version = "^2.0.5"}
tqdm = "^4.67.1"
optax = "^0.2.4"
flax = "^0.10.2"
jaxtyping = "^0.2.36"
pyarrow = "^18.1.0"
google-cloud-storage = "^2.19.0"
typer = "^0.15.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
